refactor(cli): extract shared parseApiUrl validator (review)
Dedupe the trim/strip-slash/http(s) URL validation that was copied between
the --api-url flag (apiCommand.ts) and the first-run EnvironmentPicker into a
single parseApiUrl() in utils/apiUrl.ts, its pure home. Both callers now share
one source of truth; validator tests move alongside it in apiUrl.test.ts.

Addresses self-review nit on #143.
