+++ packages/cli/src/utils/index.ts
@@ -10,6 +10,7 @@ export {
   isSourceMode,
   LOCAL_DEV_URL,
   getDefaultApiUrl,
+  parseApiUrl,
   type ApiEndpoint,
 } from './apiUrl';
 export { logger } from './Logger';
